+++ lib/app/modules/home/sub_views/view_transactions/view_transactions.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:money/app/controller/preferences_controller.dart';
 import 'package:money/app/core/helpers/date_helper.dart';
 import 'package:money/app/core/helpers/list_helper.dart';
+import 'package:money/app/core/helpers/misc_helpers.dart';
 import 'package:money/app/core/helpers/ranges.dart';
 import 'package:money/app/core/helpers/string_helper.dart';
 import 'package:money/app/core/widgets/columns/footer_widgets.dart';
@@ -164,6 +165,17 @@ class ViewTransactionsState extends ViewForMoneyObjectsState {
                     columnFilter: [],
                     textFilter: '',
                   );
+                }
+              },
+            ),
+            // Search Payee
+            InternalViewSwitching(
+              Icons.person_search_outlined,
+              'Search for Payee',
+              () {
+                final transaction = getFirstSelectedItem() as Transaction?;
+                if (transaction != null) {
+                  launchGoogleSearch(transaction.getPayeeOrTransferCaption());
                 }
               },
             ),